+++ R/shinyalert.R
@@ -5,9 +5,7 @@
 #' and you can specify custom code to run when a modal closes. See the
 #' \href{https://daattali.com/shiny/shinyalert-demo/}{demo Shiny app}
 #' online for examples or read the
-#' \href{https://github.com/daattali/shinyalert#readme}{full README}.\cr\cr
-#' \code{shinyalert} must be initialized with a call to
-#' \code{\link[shinyalert]{useShinyalert}} in the app's UI.
+#' \href{https://github.com/daattali/shinyalert#readme}{full README}.
 #'
 #' @param title The title of the modal.
 #' @param text The modal's text. Can either be simple text, or Shiny tags (including
@@ -62,6 +60,7 @@
 #' for small (default), `"m"` for medium, or `"l"` for large.
 #' @param immediate If `TRUE`, close any previously opened alerts and display
 #' the current one immediately.
+#' @param session Shiny session object (only for advanced users).
 #'
 #' @return An ID that can be used by \code{\link{closeAlert}} to close this
 #' specific alert.
@@ -159,7 +158,6 @@
 #'
 #'   shinyApp(
 #'     ui = fluidPage(
-#'       useShinyalert(),  # Set up shinyalert
 #'       actionButton("btn", "Click me")
 #'     ),
 #'     server = function(input, output) {
@@ -178,7 +176,6 @@
 #'
 #'   shinyApp(
 #'     ui = fluidPage(
-#'       useShinyalert(),  # Set up shinyalert
 #'       actionButton("btn", "Greet")
 #'     ),
 #'     server = function(input, output) {
@@ -199,7 +196,6 @@
 #'
 #'   shinyApp(
 #'     ui = fluidPage(
-#'       useShinyalert(),  # Set up shinyalert
 #'       actionButton("btn", "Go")
 #'     ),
 #'     server = function(input, output) {
@@ -244,7 +240,8 @@ shinyalert <- function(
   callbackJS = NULL,
   inputId = "shinyalert",
   size = "s",
-  immediate = FALSE
+  immediate = FALSE,
+  session = getSession()
 ) {
 
   params <- as.list(environment())
@@ -275,7 +272,13 @@ shinyalert <- function(
   params[['imageWidth']] <- NULL
   params[['imageHeight']] <- NULL
 
-  session <- getSession()
+  # Remove parameters that shouldn't be passed to JavaScript
+  params[['session']] <- NULL
+
+  if (is.null(session$userData$.shinyalert_added) || !session$userData$.shinyalert_added) {
+    shiny::insertUI("head", "beforeEnd", getDependencies(), immediate = TRUE)
+    session$userData$.shinyalert_added <- TRUE
+  }
 
   if (immediate) {
     closeAlert()